# pta-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/auth/parent/register": {
      "post": {
        "operationId": "AuthController_registerParent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parent registered successfully and verification code dispatched"
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new parent account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/teacher/register": {
      "post": {
        "operationId": "AuthController_registerTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teacher registered successfully and verification code dispatched"
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new teacher account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired verification code"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Verify email using 6-digit code delivered via Brevo",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/resend-code": {
      "post": {
        "operationId": "AuthController_resendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code resent if account exists"
          },
          "400": {
            "description": "Rate limit or already verified"
          }
        },
        "summary": "Resend 6-digit verification code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/parent/login": {
      "post": {
        "operationId": "AuthController_loginParent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns tokens and user"
          },
          "401": {
            "description": "Invalid credentials or role mismatch"
          },
          "403": {
            "description": "Account pending verification, suspended, or disabled"
          }
        },
        "summary": "Log in as a parent",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/teacher/login": {
      "post": {
        "operationId": "AuthController_loginTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns tokens and user"
          },
          "401": {
            "description": "Invalid credentials or role mismatch"
          },
          "403": {
            "description": "Account pending approval, suspended, or disabled"
          }
        },
        "summary": "Log in as a teacher",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token rotated successfully"
          },
          "401": {
            "description": "Invalid, expired, or reused refresh token"
          }
        },
        "summary": "Rotate and exchange a refresh token for new tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out and invalidate active session tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset code sent if account exists"
          },
          "400": {
            "description": "Validation failed or rate limited"
          }
        },
        "summary": "Initiate password reset flow by sending a 6-digit OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid, expired, or max-attempt exceeded reset code"
          }
        },
        "summary": "Reset account password using 6-digit OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid new password"
          },
          "401": {
            "description": "Current password mismatch or unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password for currently authenticated user",
        "tags": [
          "Authentication"
        ]
      }
    }
  },
  "info": {
    "title": "PTA Authentication API",
    "description": "Authentication Module with Role-Based Access Control and Brevo OTP Verification",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "RegisterParentDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "parent@example.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssword123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "example": "Sarah Johnson"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "RegisterTeacherDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "teacher@example.com"
          },
          "password": {
            "type": "string",
            "example": "T@acherPass123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "example": "Mr. David Clark"
          },
          "phone": {
            "type": "string",
            "example": "+1234567891"
          },
          "subjectSpecialization": {
            "type": "string",
            "example": "Mathematics & Science"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Six-digit verification code"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResendCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Hashed or random refresh token string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Six-digit password reset verification code"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecurePassword123!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "CurrentPassword123!"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecurePassword123!",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      }
    }
  }
}